+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Task filters now properly reset when navigating between projects
+
 ## [0.19.0] - 2026-01-30
 
 ### Added